// CUI-OTA/src/app/models/base.model.ts
export enum IENMessage {
    resetCashingSuccess = 'reset cashing success',
    resetCashingFail = 'reset cashing fail',
    verifyOrderFail = 'verify order fail',
    sumerizeUnmatch = 'sumerize unmatch',
    sumerizeOrdersEmpty = 'sumerize orders empty',
    vendingPendingSum = 'vending_pending_sum',
    ordersUnmatch = 'orders unmatch',
    orderEmpty = 'order empty',
    vendingPendingOrders = 'vending_pending_order',
    orderCanceled = 'order canceled',
    invalidSumerizeOrder = 'invalid sumerize order',
    installVendingVersion = 'install vending version',
    downloadingVendingVersion = 'download vending version',
    AppConfig = 'App Config',
    installingNewVersionFail = 'installing new version fail',
    installingNewVersionSuccess = 'installing new version success',
    extractingFilePack3 = 'extracting file pack_3',
    extractingFilePack2 = 'extracting file pack_2',
    systemIsRepairingPleaseWaitMoment = 'System is repairing please wait moment',
    installingNewVersion = 'installing new version',
    extractingFile = 'extracting file',
    pending = 'pending',
    repairFail = 'repair fail',
    repairSystemComplete = 'repaire system complete',
    qrcodeExpired = 'QR Code expired',
    timeout = 'Time out !',
    invalidSellDetail = 'invalid sell detail',
    retryPayYourOrderFail5TimesYourOrderWasReset = 'retry pay your order fail 5 times your order was reset',
    generateMMoneyQRCodeFail = 'generate MMoney QRCode fail',
    invalidAmount = 'invalid amount',
    invalidSelectionGameMenu = 'invalid selection game menu',
    notFoundYourMMoneyAccount = 'Not found your MMoney account',
    saveSaleFail = 'save sale fail',
    endShouldMoreThenStart = 'end should more then start',
    startShouldLesterThenEnd = 'start should lester then end',
    cashoutMMoneySuccess = 'cash out MMoney success',
    testMotor = 'test motor',
    testMotorSuccess = 'test motors success',
    notFoundFile = 'not found file',
    vendingSaleListEmpty = 'vending sale list empty',
    loadVendingSaleListFail = 'load vending sale list fail',
    phonePaymentSuccess = 'phone payment success',
    minimumOfAmountIs1000 = 'minimum of amount is 1000',
    invalidCustomAmount = 'invalid custom amount',
    balanceIsNotEnought = 'balance is not enought',
    invalidSelectionTopupServiceMenu = 'invalid selection topup service menu',
    invalidPhonenumber = 'invalid phone number',
    thereIsNotBalance = 'there is not balance',
    cashoutToAnotherLAABAccountSuccess = 'cash out to another LAAB account success',
    laabCashinSuccess = 'LAAB cash in success',
    timeupPleaseGenerateAgain = 'time up please generate again',
    notFoundAnySaveSMC = 'not found any save smc',
    thisSmcHaveAlreadyUsedOrThatOneWasDelete = 'this smart contract already used or that one was delete',
    notEnoughtCashBalance = 'not enought balance',
    success = 'success',
    parametersEmpty = 'parameters empty',
    findLaoQRPaidFail = "findLaoQRPaidFail"
}

export enum ITabVendingSegement {
    vending = 'vending',
    hangmistore = 'hangmistore',
    hangmifood = 'hangmifood',
    topupandservices = 'topupandservices'
}
export enum IWebviewTabs {
    vending = 'vending',
    hangmistore = 'hangmistore',
    hangmifood = 'hangmifood',
    topupandservices = 'topupandservices'
}

export enum IGenerateQR {
    mmoney = 'mmoney',
}